Find head if album contains only sub-albums
Fixes issue where header image is set to empty album icon if album
contains only subalbums. Updated to recursively search for subalbums
with cover image and update if needed, and then update parent album.
Parent album image is set to empty only if there are no images, and all
subalbums are also empty.

This does not take into account whether trip reports are present. May
want to update this in the future.